src/main/java/lotto/domain/Lottos.java
package lotto.domain;

import java.util.List;
import java.util.stream.Collectors;
import java.util.stream.Stream;

public class Lottos {
    private final List<Lotto> lottos;

    public Lottos(List<Lotto> lottos) {
        this.lottos = lottos;
    }

    public static Lottos combine(Lottos one, Lottos another) {
        return new Lottos(
                Stream.concat(one.lottos.stream(), another.lottos.stream())
                        .collect(Collectors.toList()));
    }

    public int getSize() {
        return lottos.size();
    }

    public List<Lotto> getLottos() {
        return lottos;
    }
}
